Add HTTP client CLI for making requests with methods,headers and data
